Split render pipleline
