+++ crates/bevy_sprite/src/render/draw.rs
@@ -1,0 +1,87 @@
+use super::{ImageBindGroups, SpriteBatch, SpriteMeta};
+use bevy_asset::Handle;
+use bevy_ecs::prelude::*;
+use bevy_ecs::system::{
+    lifetimeless::{Read, SQuery, SRes},
+    SystemParamItem,
+};
+use bevy_render::{
+    render_phase::{
+        BatchedPhaseItem, EntityRenderCommand, RenderCommand, RenderCommandResult, SetItemPipeline,
+        TrackedRenderPass,
+    },
+    view::ViewUniformOffset,
+};
+
+pub type DrawSprite = (
+    SetItemPipeline,
+    SetSpriteViewBindGroup<0>,
+    SetSpriteTextureBindGroup<1>,
+    DrawSpriteBatch,
+);
+
+pub struct SetSpriteViewBindGroup<const I: usize>;
+impl<const I: usize> EntityRenderCommand for SetSpriteViewBindGroup<I> {
+    type Param = (SRes<SpriteMeta>, SQuery<Read<ViewUniformOffset>>);
+
+    fn render<'w>(
+        view: Entity,
+        _item: Entity,
+        (sprite_meta, view_query): SystemParamItem<'w, '_, Self::Param>,
+        pass: &mut TrackedRenderPass<'w>,
+    ) -> RenderCommandResult {
+        let view_uniform = view_query.get(view).unwrap();
+        pass.set_bind_group(
+            I,
+            sprite_meta.into_inner().view_bind_group.as_ref().unwrap(),
+            &[view_uniform.offset],
+        );
+        RenderCommandResult::Success
+    }
+}
+pub struct SetSpriteTextureBindGroup<const I: usize>;
+impl<const I: usize> EntityRenderCommand for SetSpriteTextureBindGroup<I> {
+    type Param = (SRes<ImageBindGroups>, SQuery<Read<SpriteBatch>>);
+
+    fn render<'w>(
+        _view: Entity,
+        item: Entity,
+        (image_bind_groups, query_batch): SystemParamItem<'w, '_, Self::Param>,
+        pass: &mut TrackedRenderPass<'w>,
+    ) -> RenderCommandResult {
+        let sprite_batch = query_batch.get(item).unwrap();
+        let image_bind_groups = image_bind_groups.into_inner();
+
+        pass.set_bind_group(
+            I,
+            image_bind_groups
+                .values
+                .get(&Handle::weak(sprite_batch.image_handle_id))
+                .unwrap(),
+            &[],
+        );
+        RenderCommandResult::Success
+    }
+}
+
+pub struct DrawSpriteBatch;
+impl<P: BatchedPhaseItem> RenderCommand<P> for DrawSpriteBatch {
+    type Param = (SRes<SpriteMeta>, SQuery<Read<SpriteBatch>>);
+
+    fn render<'w>(
+        _view: Entity,
+        item: &P,
+        (sprite_meta, query_batch): SystemParamItem<'w, '_, Self::Param>,
+        pass: &mut TrackedRenderPass<'w>,
+    ) -> RenderCommandResult {
+        let sprite_batch = query_batch.get(item.entity()).unwrap();
+        let sprite_meta = sprite_meta.into_inner();
+        if sprite_batch.colored {
+            pass.set_vertex_buffer(0, sprite_meta.colored_vertices.buffer().unwrap().slice(..));
+        } else {
+            pass.set_vertex_buffer(0, sprite_meta.vertices.buffer().unwrap().slice(..));
+        }
+        pass.draw(item.batch_range().as_ref().unwrap().clone(), 0..1);
+        RenderCommandResult::Success
+    }
+}
