+++ crates/bevy_sprite/src/render/mod.rs
@@ -1,40 +1,24 @@
-use std::cmp::Ordering;
+pub mod draw;
+pub mod extract;
+pub mod prepare;
+pub mod queue;
 
-use crate::{
-    texture_atlas::{TextureAtlas, TextureAtlasSprite},
-    Sprite, SPRITE_SHADER_HANDLE,
-};
-use bevy_asset::{AssetEvent, Assets, Handle, HandleId};
-use bevy_core_pipeline::{core_2d::Transparent2d, tonemapping::Tonemapping};
-use bevy_ecs::{
-    prelude::*,
-    system::{lifetimeless::*, SystemParamItem, SystemState},
-};
-use bevy_math::{Rect, Vec2};
-use bevy_reflect::Uuid;
+pub use draw::DrawSprite;
+pub use extract::{ExtractedSprite, ExtractedSprites, SpriteAssetEvents};
+pub use prepare::{PreparedSprite, PreparedSprites};
+pub use queue::{ImageBindGroups, SpriteBatch, SpriteMeta};
+
+use crate::SPRITE_SHADER_HANDLE;
+use bevy_ecs::{prelude::*, system::SystemState};
+use bevy_math::Vec2;
 use bevy_render::{
-    color::Color,
-    render_asset::RenderAssets,
-    render_phase::{
-        BatchedPhaseItem, DrawFunctions, EntityRenderCommand, RenderCommand, RenderCommandResult,
-        RenderPhase, SetItemPipeline, TrackedRenderPass,
-    },
     render_resource::*,
     renderer::{RenderDevice, RenderQueue},
     texture::{
         BevyDefault, DefaultImageSampler, GpuImage, Image, ImageSampler, TextureFormatPixelInfo,
     },
-    view::{
-        ComputedVisibility, ExtractedView, Msaa, ViewTarget, ViewUniform, ViewUniformOffset,
-        ViewUniforms, VisibleEntities,
-    },
-    Extract,
+    view::{ViewTarget, ViewUniform},
 };
-use bevy_transform::components::GlobalTransform;
-use bevy_utils::FloatOrd;
-use bevy_utils::HashMap;
-use bytemuck::{Pod, Zeroable};
-use fixedbitset::FixedBitSet;
 
 #[derive(Resource)]
 pub struct SpritePipeline {
@@ -260,528 +244,5 @@ impl SpecializedRenderPipeline for SpritePipeline {
             },
             label: Some("sprite_pipeline".into()),
         }
-    }
-}
-
-#[derive(Clone, Copy)]
-pub struct ExtractedSprite {
-    pub entity: Entity,
-    pub transform: GlobalTransform,
-    pub color: Color,
-    /// Select an area of the texture
-    pub rect: Option<Rect>,
-    /// Change the on-screen size of the sprite
-    pub custom_size: Option<Vec2>,
-    /// Handle to the `Image` of this sprite
-    /// PERF: storing a `HandleId` instead of `Handle<Image>` enables some optimizations (`ExtractedSprite` becomes `Copy` and doesn't need to be dropped)
-    pub image_handle_id: HandleId,
-    pub flip_x: bool,
-    pub flip_y: bool,
-    pub anchor: Vec2,
-}
-
-#[derive(Resource, Default)]
-pub struct ExtractedSprites {
-    pub sprites: Vec<ExtractedSprite>,
-}
-
-#[derive(Resource, Default)]
-pub struct SpriteAssetEvents {
-    pub images: Vec<AssetEvent<Image>>,
-}
-
-pub fn extract_sprite_events(
-    mut events: ResMut<SpriteAssetEvents>,
-    mut image_events: Extract<EventReader<AssetEvent<Image>>>,
-) {
-    let SpriteAssetEvents { ref mut images } = *events;
-    images.clear();
-
-    for image in image_events.iter() {
-        // AssetEvent: !Clone
-        images.push(match image {
-            AssetEvent::Created { handle } => AssetEvent::Created {
-                handle: handle.clone_weak(),
-            },
-            AssetEvent::Modified { handle } => AssetEvent::Modified {
-                handle: handle.clone_weak(),
-            },
-            AssetEvent::Removed { handle } => AssetEvent::Removed {
-                handle: handle.clone_weak(),
-            },
-        });
-    }
-}
-
-pub fn extract_sprites(
-    mut extracted_sprites: ResMut<ExtractedSprites>,
-    texture_atlases: Extract<Res<Assets<TextureAtlas>>>,
-    sprite_query: Extract<
-        Query<(
-            Entity,
-            &ComputedVisibility,
-            &Sprite,
-            &GlobalTransform,
-            &Handle<Image>,
-        )>,
-    >,
-    atlas_query: Extract<
-        Query<(
-            Entity,
-            &ComputedVisibility,
-            &TextureAtlasSprite,
-            &GlobalTransform,
-            &Handle<TextureAtlas>,
-        )>,
-    >,
-) {
-    for (entity, visibility, sprite, transform, handle) in sprite_query.iter() {
-        if !visibility.is_visible() {
-            continue;
-        }
-        // PERF: we don't check in this function that the `Image` asset is ready, since it should be in most cases and hashing the handle is expensive
-        extracted_sprites.sprites.push(ExtractedSprite {
-            entity,
-            color: sprite.color,
-            transform: *transform,
-            rect: sprite.rect,
-            // Pass the custom size
-            custom_size: sprite.custom_size,
-            flip_x: sprite.flip_x,
-            flip_y: sprite.flip_y,
-            image_handle_id: handle.id(),
-            anchor: sprite.anchor.as_vec(),
-        });
-    }
-    for (entity, visibility, atlas_sprite, transform, texture_atlas_handle) in atlas_query.iter() {
-        if !visibility.is_visible() {
-            continue;
-        }
-        if let Some(texture_atlas) = texture_atlases.get(texture_atlas_handle) {
-            let rect = Some(texture_atlas.textures[atlas_sprite.index]);
-            extracted_sprites.sprites.push(ExtractedSprite {
-                entity,
-                color: atlas_sprite.color,
-                transform: *transform,
-                // Select the area in the texture atlas
-                rect,
-                // Pass the custom size
-                custom_size: atlas_sprite.custom_size,
-                flip_x: atlas_sprite.flip_x,
-                flip_y: atlas_sprite.flip_y,
-                image_handle_id: texture_atlas.texture.id(),
-                anchor: atlas_sprite.anchor.as_vec(),
-            });
-        }
-    }
-}
-
-const QUAD_VERTEX_POSITIONS: [Vec2; 4] = [
-    Vec2::new(-0.5, -0.5),
-    Vec2::new(0.5, -0.5),
-    Vec2::new(0.5, 0.5),
-    Vec2::new(-0.5, 0.5),
-];
-
-const QUAD_UVS: [Vec2; 4] = [
-    Vec2::new(0., 1.),
-    Vec2::new(1., 1.),
-    Vec2::new(1., 0.),
-    Vec2::new(0., 0.),
-];
-
-pub struct PreparedSprite {
-    /// The main world associated entity
-    entity: Entity,
-    vertex_positions: [[f32; 3]; 4],
-    vertex_uvs: [[f32; 2]; 4],
-    image_handle_id: HandleId,
-    color: [f32; 4],
-    sort_key: f32,
-}
-
-#[derive(Resource, Default)]
-pub struct PreparedSprites {
-    pub sprites: Vec<PreparedSprite>,
-}
-
-pub fn prepare_sprites(
-    mut extracted_sprites: ResMut<ExtractedSprites>,
-    mut prepared_sprites: ResMut<PreparedSprites>,
-    gpu_images: Res<RenderAssets<Image>>,
-) {
-    prepared_sprites.sprites.clear();
-    for extracted_sprite in extracted_sprites.sprites.drain(..) {
-        let current_image_size =
-            match gpu_images.get(&Handle::weak(extracted_sprite.image_handle_id)) {
-                None => continue,
-                Some(img) => img.size,
-            };
-
-        // Calculate vertex data for this item
-
-        let mut uvs = QUAD_UVS;
-        if extracted_sprite.flip_x {
-            uvs = [uvs[1], uvs[0], uvs[3], uvs[2]];
-        }
-        if extracted_sprite.flip_y {
-            uvs = [uvs[3], uvs[2], uvs[1], uvs[0]];
-        }
-
-        // By default, the size of the quad is the size of the texture
-        let mut quad_size = current_image_size;
-
-        // If a rect is specified, adjust UVs and the size of the quad
-        if let Some(rect) = extracted_sprite.rect {
-            let rect_size = rect.size();
-            for uv in &mut uvs {
-                *uv = (rect.min + *uv * rect_size) / current_image_size;
-            }
-            quad_size = rect_size;
-        }
-
-        // Override the size if a custom one is specified
-        if let Some(custom_size) = extracted_sprite.custom_size {
-            quad_size = custom_size;
-        }
-
-        // Apply size and global transform
-        let vertex_positions = QUAD_VERTEX_POSITIONS.map(|quad_pos| {
-            extracted_sprite
-                .transform
-                .transform_point(((quad_pos - extracted_sprite.anchor) * quad_size).extend(0.))
-                .into()
-        });
-
-        let sort_key = extracted_sprite.transform.translation().z;
-
-        prepared_sprites.sprites.push(PreparedSprite {
-            entity: extracted_sprite.entity,
-            vertex_positions,
-            vertex_uvs: uvs.map(Into::into),
-            color: extracted_sprite.color.as_linear_rgba_f32(),
-            image_handle_id: extracted_sprite.image_handle_id,
-            sort_key,
-        });
-    }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-struct SpriteVertex {
-    pub position: [f32; 3],
-    pub uv: [f32; 2],
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-struct ColoredSpriteVertex {
-    pub position: [f32; 3],
-    pub uv: [f32; 2],
-    pub color: [f32; 4],
-}
-
-#[derive(Resource)]
-pub struct SpriteMeta {
-    vertices: BufferVec<SpriteVertex>,
-    colored_vertices: BufferVec<ColoredSpriteVertex>,
-    view_bind_group: Option<BindGroup>,
-}
-
-impl Default for SpriteMeta {
-    fn default() -> Self {
-        Self {
-            vertices: BufferVec::new(BufferUsages::VERTEX),
-            colored_vertices: BufferVec::new(BufferUsages::VERTEX),
-            view_bind_group: None,
-        }
-    }
-}
-
-const QUAD_INDICES: [usize; 6] = [0, 2, 3, 0, 1, 2];
-
-#[derive(Component, Eq, PartialEq, Copy, Clone)]
-pub struct SpriteBatch {
-    image_handle_id: HandleId,
-    colored: bool,
-}
-
-#[derive(Resource, Default)]
-pub struct ImageBindGroups {
-    values: HashMap<Handle<Image>, BindGroup>,
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn queue_sprites(
-    mut commands: Commands,
-    mut view_entities: Local<FixedBitSet>,
-    draw_functions: Res<DrawFunctions<Transparent2d>>,
-    render_device: Res<RenderDevice>,
-    render_queue: Res<RenderQueue>,
-    mut sprite_meta: ResMut<SpriteMeta>,
-    view_uniforms: Res<ViewUniforms>,
-    sprite_pipeline: Res<SpritePipeline>,
-    mut pipelines: ResMut<SpecializedRenderPipelines<SpritePipeline>>,
-    mut pipeline_cache: ResMut<PipelineCache>,
-    mut image_bind_groups: ResMut<ImageBindGroups>,
-    gpu_images: Res<RenderAssets<Image>>,
-    msaa: Res<Msaa>,
-    mut prepared_sprites: ResMut<PreparedSprites>,
-    mut views: Query<(
-        &mut RenderPhase<Transparent2d>,
-        &VisibleEntities,
-        &ExtractedView,
-        Option<&Tonemapping>,
-    )>,
-    events: Res<SpriteAssetEvents>,
-) {
-    // If an image has changed, the GpuImage has (probably) changed
-    for event in &events.images {
-        match event {
-            AssetEvent::Created { .. } => None,
-            AssetEvent::Modified { handle } | AssetEvent::Removed { handle } => {
-                image_bind_groups.values.remove(handle)
-            }
-        };
-    }
-
-    let msaa_key = SpritePipelineKey::from_msaa_samples(msaa.samples);
-
-    if let Some(view_binding) = view_uniforms.uniforms.binding() {
-        let sprite_meta = &mut sprite_meta;
-
-        // Clear the vertex buffers
-        sprite_meta.vertices.clear();
-        sprite_meta.colored_vertices.clear();
-
-        sprite_meta.view_bind_group = Some(render_device.create_bind_group(&BindGroupDescriptor {
-            entries: &[BindGroupEntry {
-                binding: 0,
-                resource: view_binding,
-            }],
-            label: Some("sprite_view_bind_group"),
-            layout: &sprite_pipeline.view_layout,
-        }));
-
-        let draw_sprite_function = draw_functions.read().id::<DrawSprite>();
-
-        // Vertex buffer indices
-        let mut index = 0;
-        let mut colored_index = 0;
-
-        // FIXME: VisibleEntities is ignored
-
-        let prepared_sprites = &mut prepared_sprites.sprites;
-        // Sort sprites by z for correct transparency and then by handle to improve batching
-        // NOTE: This can be done independent of views by reasonably assuming that all 2D views look along the negative-z axis in world space
-        prepared_sprites.sort_unstable_by(|a, b| match a.sort_key.partial_cmp(&b.sort_key) {
-            Some(Ordering::Equal) | None => a.image_handle_id.cmp(&b.image_handle_id),
-            Some(other) => other,
-        });
-        let image_bind_groups = &mut *image_bind_groups;
-
-        for (mut transparent_phase, visible_entities, view, tonemapping) in &mut views {
-            let mut view_key = SpritePipelineKey::from_hdr(view.hdr) | msaa_key;
-            if let Some(Tonemapping::Enabled { deband_dither }) = tonemapping {
-                if !view.hdr {
-                    view_key |= SpritePipelineKey::TONEMAP_IN_SHADER;
-
-                    if *deband_dither {
-                        view_key |= SpritePipelineKey::DEBAND_DITHER;
-                    }
-                }
-            }
-            let pipeline = pipelines.specialize(
-                &mut pipeline_cache,
-                &sprite_pipeline,
-                view_key | SpritePipelineKey::from_colored(false),
-            );
-            let colored_pipeline = pipelines.specialize(
-                &mut pipeline_cache,
-                &sprite_pipeline,
-                view_key | SpritePipelineKey::from_colored(true),
-            );
-
-            view_entities.clear();
-            view_entities.extend(visible_entities.entities.iter().map(|e| e.index() as usize));
-            transparent_phase.items.reserve(prepared_sprites.len());
-
-            // Impossible starting values that will be replaced on the first iteration
-            let mut current_batch = SpriteBatch {
-                image_handle_id: HandleId::Id(Uuid::nil(), u64::MAX),
-                colored: false,
-            };
-            let mut current_batch_entity = Entity::PLACEHOLDER;
-            // Add a phase item for each sprite, and detect when succesive items can be batched.
-            // Spawn an entity with a `SpriteBatch` component for each possible batch.
-            // Compatible items share the same entity.
-            // Batches are merged later (in `batch_phase_system()`), so that they can be interrupted
-            // by any other phase item (and they can interrupt other items from batching).
-            for prepared_sprite in prepared_sprites.iter() {
-                if !view_entities.contains(prepared_sprite.entity.index() as usize) {
-                    continue;
-                }
-                let new_batch = SpriteBatch {
-                    image_handle_id: prepared_sprite.image_handle_id,
-                    colored: prepared_sprite.color != [1.0; 4],
-                };
-                if new_batch != current_batch {
-                    // Set-up a new possible batch
-                    if let Some(gpu_image) =
-                        gpu_images.get(&Handle::weak(new_batch.image_handle_id))
-                    {
-                        current_batch = new_batch;
-                        current_batch_entity = commands.spawn(current_batch).id();
-
-                        image_bind_groups
-                            .values
-                            .entry(Handle::weak(current_batch.image_handle_id))
-                            .or_insert_with(|| {
-                                render_device.create_bind_group(&BindGroupDescriptor {
-                                    entries: &[
-                                        BindGroupEntry {
-                                            binding: 0,
-                                            resource: BindingResource::TextureView(
-                                                &gpu_image.texture_view,
-                                            ),
-                                        },
-                                        BindGroupEntry {
-                                            binding: 1,
-                                            resource: BindingResource::Sampler(&gpu_image.sampler),
-                                        },
-                                    ],
-                                    label: Some("sprite_material_bind_group"),
-                                    layout: &sprite_pipeline.material_layout,
-                                })
-                            });
-                    } else {
-                        // Skip this item if the texture is not ready
-                        continue;
-                    }
-                }
-
-                // These items will be sorted by depth with other phase items
-                let sort_key = FloatOrd(prepared_sprite.sort_key);
-
-                // Store the vertex data and add the item to the render phase
-                if current_batch.colored {
-                    for i in QUAD_INDICES {
-                        sprite_meta.colored_vertices.push(ColoredSpriteVertex {
-                            position: prepared_sprite.vertex_positions[i],
-                            uv: prepared_sprite.vertex_uvs[i],
-                            color: prepared_sprite.color,
-                        });
-                    }
-                    let item_start = colored_index;
-                    colored_index += QUAD_INDICES.len() as u32;
-                    let item_end = colored_index;
-
-                    transparent_phase.add(Transparent2d {
-                        draw_function: draw_sprite_function,
-                        pipeline: colored_pipeline,
-                        entity: current_batch_entity,
-                        sort_key,
-                        batch_range: Some(item_start..item_end),
-                    });
-                } else {
-                    for i in QUAD_INDICES {
-                        sprite_meta.vertices.push(SpriteVertex {
-                            position: prepared_sprite.vertex_positions[i],
-                            uv: prepared_sprite.vertex_uvs[i],
-                        });
-                    }
-                    let item_start = index;
-                    index += QUAD_INDICES.len() as u32;
-                    let item_end = index;
-
-                    transparent_phase.add(Transparent2d {
-                        draw_function: draw_sprite_function,
-                        pipeline,
-                        entity: current_batch_entity,
-                        sort_key,
-                        batch_range: Some(item_start..item_end),
-                    });
-                }
-            }
-        }
-        sprite_meta
-            .vertices
-            .write_buffer(&render_device, &render_queue);
-        sprite_meta
-            .colored_vertices
-            .write_buffer(&render_device, &render_queue);
-    }
-}
-
-pub type DrawSprite = (
-    SetItemPipeline,
-    SetSpriteViewBindGroup<0>,
-    SetSpriteTextureBindGroup<1>,
-    DrawSpriteBatch,
-);
-
-pub struct SetSpriteViewBindGroup<const I: usize>;
-impl<const I: usize> EntityRenderCommand for SetSpriteViewBindGroup<I> {
-    type Param = (SRes<SpriteMeta>, SQuery<Read<ViewUniformOffset>>);
-
-    fn render<'w>(
-        view: Entity,
-        _item: Entity,
-        (sprite_meta, view_query): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let view_uniform = view_query.get(view).unwrap();
-        pass.set_bind_group(
-            I,
-            sprite_meta.into_inner().view_bind_group.as_ref().unwrap(),
-            &[view_uniform.offset],
-        );
-        RenderCommandResult::Success
-    }
-}
-pub struct SetSpriteTextureBindGroup<const I: usize>;
-impl<const I: usize> EntityRenderCommand for SetSpriteTextureBindGroup<I> {
-    type Param = (SRes<ImageBindGroups>, SQuery<Read<SpriteBatch>>);
-
-    fn render<'w>(
-        _view: Entity,
-        item: Entity,
-        (image_bind_groups, query_batch): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let sprite_batch = query_batch.get(item).unwrap();
-        let image_bind_groups = image_bind_groups.into_inner();
-
-        pass.set_bind_group(
-            I,
-            image_bind_groups
-                .values
-                .get(&Handle::weak(sprite_batch.image_handle_id))
-                .unwrap(),
-            &[],
-        );
-        RenderCommandResult::Success
-    }
-}
-
-pub struct DrawSpriteBatch;
-impl<P: BatchedPhaseItem> RenderCommand<P> for DrawSpriteBatch {
-    type Param = (SRes<SpriteMeta>, SQuery<Read<SpriteBatch>>);
-
-    fn render<'w>(
-        _view: Entity,
-        item: &P,
-        (sprite_meta, query_batch): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let sprite_batch = query_batch.get(item.entity()).unwrap();
-        let sprite_meta = sprite_meta.into_inner();
-        if sprite_batch.colored {
-            pass.set_vertex_buffer(0, sprite_meta.colored_vertices.buffer().unwrap().slice(..));
-        } else {
-            pass.set_vertex_buffer(0, sprite_meta.vertices.buffer().unwrap().slice(..));
-        }
-        pass.draw(item.batch_range().as_ref().unwrap().clone(), 0..1);
-        RenderCommandResult::Success
     }
 }
